requirements.txt
Appium-Python-Client==4.2.1
attrs==24.2.0
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==7.0.1
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0

black==23.3.0
mypy==0.991
pylint==2.15.10
